Remove superfluous calc
Currently the value of the padding-bottom property is invalid due to the fact that the calc function is output in the CSS which does not understand sass functions and variables `calc(oTechdocsSpacingUnit() - $border-thickness)`. This means that the computed value of the padding-bottom property is whatever the default value the user agent has decided to go with. Chrome goes with `0px`.

The values used in the calc are static, meaning that we can remove the calc call and let this be done by the preprocessor.

Removing the calc call surfaced the fact that the function requires an argument. The argument could be `0` but that would make the resulting value become `-4px` and negative values are not allowed on padding. When the argument is `1` then the resulting values becomes `4px`.